First commit
Need to pass a keystore to create a signed apk
